refactor(ui): 把 DangerZone 拆出独立组件，移到关于页最末尾
「危险操作 / 一键卸载」原本在 EnvironmentDoctorPanel 内，紧贴在「环境
状态」下方，跟下面的「本地环境检查」「一键安装 Claude Code」高频区
块视觉相邻，存在误触风险。

拆分后：
- 新增 DangerZoneCard.tsx：自包含的卸载流程组件（流式日志 / 预览弹
  窗 / 二次确认 / 卸载报告 / 备份路径复制全部内聚），不需要任何 props
- EnvironmentDoctorPanel.tsx 移除所有 uninstall 相关 state、handler、
  Dialog 与 imports，只剩诊断展示 + 一键安装/一键修复按钮 + 安装日志
- AboutSection.tsx 在末尾追加 <DangerZoneCard />

最终关于页顺序：CC Doctor 版本卡 → 环境状态 → 本地环境检查 →
一键安装 Claude Code → 危险操作。
